# [PERF] account: speed up `_fetch_duplicate_reference`