@robodoo 2024-10-31
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/185975.png)](https://mergebot.odoo.com/odoo/odoo/pull/185975)