@pivi-odoo 2024-11-13
Tnx for the input @william-andre, I've applied your feedback, should be good to go.